% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the list of cubes from the URL server}
\usage{
describe_coverage(URL, name, .print = TRUE)
}
\arguments{
\item{URL}{URL of the server}

\item{name}{name of coverage}

\item{.print}{Print the coverage description}
}
\value{
tibble with coverage description
}
\description{
Contacts the WTSS server to describe one coverage
}
\examples{
describe_coverage("http://www.esensing.dpi.inpe.br/wtss", "MOD13Q1")
}
