% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-utils.R
\name{.parse_json}
\alias{.parse_json}
\title{Parse a JSON response from the WTSS server}
\usage{
.parse_json(response)
}
\arguments{
\item{response}{a response httr object}
}
\value{
parsed JSON document
}
\description{
Parse a JSON response from the WTSS service
}
