% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_coverage.R
\name{.rm_trailing_dash}
\alias{.rm_trailing_dash}
\title{Remove trailing dashes from a WTSS server address}
\usage{
.rm_trailing_dash(url)
}
\arguments{
\item{url}{A WTSS URL}
}
\value{
URL without trailing dash
}
\description{
The WTSS URL cannot have a trailing dash. This functions checks
and removes it, if present.
}
