% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_utils.R
\name{.wtss_align_dates}
\alias{.wtss_align_dates}
\title{Aligns dates of time series to a reference date}
\usage{
.wtss_align_dates(data, ref_dates)
}
\arguments{
\item{data}{Input tibble (useful for chaining functions).}

\item{ref_dates}{Dates to align the time series.}
}
\value{
A tibble with the converted dates (useful for chaining functions).
}
\description{
Converts the time indexes of a set of tibble to a single reference year.
This function is useful to join many time series from different years to a single year,
which is required by methods that combine many time series, such as clustering methods.
The reference year is taken from the date of the start of the time series
available in the data cube.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
