% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_utils.R
\name{.wtss_bands}
\alias{.wtss_bands}
\title{retrieve the bands associated to a time series}
\usage{
.wtss_bands(ts)
}
\arguments{
\item{ts}{Tibble with time series}
}
\value{
List of bands
}
\description{
retrieve the bands associated to a time series
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
