% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_config.R
\name{.wtss_config_satellite}
\alias{.wtss_config_satellite}
\title{retrieve the satellite associated to a given cube}
\usage{
.wtss_config_satellite(name, crs)
}
\arguments{
\item{name}{Name of the coverage}

\item{crs}{Projection crs information for the coverage}
}
\value{
List of providers associated to a service
}
\description{
retrieve the satellite associated to a given cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
