% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_config.R
\name{.wtss_config_sensor}
\alias{.wtss_config_sensor}
\title{retrieve the sensor associated to a data cube}
\usage{
.wtss_config_sensor(name, satellite)
}
\arguments{
\item{name}{Name of the coverage}

\item{satellite}{Name of the satellite}
}
\value{
List of providers associated to a service
}
\description{
retrieve the sensor associated to a data cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
