% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_config.R
\name{.wtss_guess_satellite}
\alias{.wtss_guess_satellite}
\title{Try a best guess for the type of sensor/satellite}
\usage{
.wtss_guess_satellite(crs)
}
\arguments{
\item{crs}{CRS of the coverage}
}
\value{
Name of the satellite .
}
\description{
Based on the projection, tries to guess what is the satellite.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
