% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_request.R
\name{.wtss_parse_json}
\alias{.wtss_parse_json}
\title{Parse a JSON response from the WTSS server}
\usage{
.wtss_parse_json(response)
}
\arguments{
\item{response}{valid JSON response from the WTSS service}
}
\value{
parsed JSON document
}
\description{
Parse a JSON response from the WTSS service
}
