% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_request.R
\name{.wtss_process_request}
\alias{.wtss_process_request}
\title{Process a request to the WTSS server}
\usage{
.wtss_process_request(request)
}
\arguments{
\item{request}{valid request to the WTSS service}
}
\value{
parsed JSON document
}
\description{
Process a request
}
