% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_request.R
\name{.wtss_send_request}
\alias{.wtss_send_request}
\title{Send a request to WTSS server}
\usage{
.wtss_send_request(request, ...)
}
\arguments{
\item{request}{valid request according to the WTSS protocol}

\item{...}{additional parameters that can be added in httr.}
}
\value{
response from the server
}
\description{
Sends a request to the WTSS server and times out after 10 tries
}
