% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_tibble.R
\name{.wtss_tibble}
\alias{.wtss_tibble}
\title{Create a sits tibble to store the time series information}
\usage{
.wtss_tibble()
}
\value{
A tibble.
}
\description{
This function returns an empty tibble that
contains the satellite image time series and its metadata. The columns are
<longitude, latitude, start_date, end_date, label, cube, time_series>.
WTSS functions produce a tibble as output.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
