% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_time_series.R
\name{.wtss_time_series_processing}
\alias{.wtss_time_series_processing}
\title{Processing a Time Series Result from WTSS}
\usage{
.wtss_time_series_processing(items)
}
\arguments{
\item{items}{Items retrieved from WTSS server}
}
\value{
tibble with a time series
}
\description{
Processing a Time Series Result from WTSS
}
