% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_tibble.R
\name{.wtss_to_tibble}
\alias{.wtss_to_tibble}
\title{Import time series in the zoo format to a tibble}
\usage{
.wtss_to_tibble(
  ts,
  name,
  bands,
  longitude,
  latitude,
  start_date,
  end_date,
  cov_desc
)
}
\arguments{
\item{ts}{list of time series retrieved by WTSS}

\item{name}{Name of the coverage where data comes from.}

\item{bands}{Bands to be retrieved from the time series.}

\item{longitude}{Longitude of the chosen location.}

\item{latitude}{Latitude of the chosen location.}

\item{start_date}{Starting date of the time series}

\item{end_date}{End date of the time series}

\item{cov_desc}{Description of the WTSS coverage}
}
\value{
Time series in sits tibble format.
}
\description{
Converts data from an instance of a zoo series to a sits tibble.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
