% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series-api.R
\name{fetch_timeseries}
\alias{fetch_timeseries}
\title{Retrieves all time series from every page}
\usage{
fetch_timeseries(timeseries)
}
\arguments{
\item{timeseries}{a \code{timeseries} object returned in the 
\code{time_series()} function.}
}
\value{
Time series in a tibble format.
}
\description{
Performs the pagination of all time series in the WTSS service.
This operation may take some time depending on the specified paging period.
}
\examples{
\dontrun{
# connect to a WTSS server
wtss_server <- "https://brazildatacube.dpi.inpe.br/dev/wtss/v2/"
# Retrieve a time series for a list
polygon <- list(
  type = "Polygon",
  coordinates = list(
    matrix(c(-46.27916, -13.22281,
             -46.27816, -13.22281,
             -46.27816, -13.22181,
             -46.27916, -13.22181,
             -46.27916, -13.22281),
           ncol = 2, byrow = TRUE)
  )
)
# Get the time series
modis_ts <- Rwtss::time_series(
               URL = wtss_server,
               name = "MOD13Q1-6",
               attributes = c("NDVI", "EVI", "red_reflectance"),
               geom = polygon,
               start_date = "2018-01-01T00:00:00Z",
               end_date = "2020-02-01T00:00:00Z",
               scale = TRUE,
               pagination = "P1M",
               token = token
)
# Get the next period of the time series
all_ts <- Rwtss::fetch_timeseries(modis_ts)
}
}
