% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi_ts}
\alias{ndvi_ts}
\title{Example time series from MOD13Q1 product.}
\format{
A \code{wtss} tibble with 388 samples. A \code{wtss} tibble contains 
data and metadata. The first six columns contain the metadata: satellite, 
sensor, spatial and temporal information, and the coverage from where the 
data has been extracted. The spatial location is given in longitude and 
latitude coordinates for the "WGS84" ellipsoid. The `time_series` column 
contains the time series data for each spatiotemporal location.
}
\usage{
data("ndvi_ts")
}
\description{
A dataset containing a wtss tibble, with extracted time series.
}
\keyword{datasets}
