% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{show}
\alias{show}
\title{Shows a WTSS object}
\usage{
show(wtss.obj)
}
\arguments{
\item{wtss.obj}{WTSS object}
}
\description{
Displays information about the WTSS service
}
\examples{
{
wtss <-  WTSS("http://www.esensing.dpi.inpe.br/wtss/")
wtss
}
}
