% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{summarize}
\alias{summarize}
\title{Gets the aggregated time series for an geometry.}
\usage{
summarize(
  URL,
  name,
  geom = NULL,
  attributes = NULL,
  aggregations = NULL,
  start_date = NULL,
  end_date = NULL,
  scale = FALSE,
  pixel_strategy = "center",
  qa_values = NULL,
  masked = NULL,
  token = NULL,
  ...
)
}
\arguments{
\item{URL}{URL of the server}

\item{name}{Coverage name.}

\item{attributes}{Vector of band names.}

\item{start_date}{Start date in the format yyyy-mm-dd or yyyy-mm 
depending on the coverage.}

\item{end_date}{End date in the format yyyy-mm-dd or yyyy-mm 
depending on the coverage.}

\item{token}{A character with token to be add in URL.}

\item{...}{Additional parameters that can be added in httr.}

\item{longitude}{Longitude in WGS84 coordinate system.}

\item{latitude}{Latitude in WGS84 coordinate system.}
}
\value{
time series in a tibble format (NULL)
}
\description{
...
}
\examples{
\dontrun{
# connect to a WTSS server
wtss_server <- "https://brazildatacube.dpi.inpe.br/wtss/"
# retrieve a time series
ndvi_ts <- Rwtss::time_series(
               URL = wtss_server, 
               name = "LC8_30_16D_STK-1", 
               attributes = "NDVI", 
               latitude = -14.31, 
               longitude = -51.16,
               token = "change-me"
)
# plot the time series
plot(ndvi_ts)
}
}
