% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_time_series.R
\name{time_series}
\alias{time_series}
\title{Get time series}
\usage{
time_series(
  wtss.obj,
  name,
  attributes = NULL,
  longitude,
  latitude,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{wtss.obj}{Either a WTSS object or a server URL}

\item{name}{Coverage name.}

\item{attributes}{Vector of band names.}

\item{longitude}{Longitude in WGS84 coordinate system.}

\item{latitude}{Latitude in WGS84 coordinate system.}

\item{start_date}{Start date in the format yyyy-mm-dd or yyyy-mm 
depending on the coverage.}

\item{end_date}{End date in the format yyyy-mm-dd or yyyy-mm 
depending on the coverage.}
}
\value{
time series in a tibble format (NULL )
}
\description{
Retrieves the time series for a pair of coordinates
}
\examples{
\dontrun{
# connect to a WTSS server
wtss_server <- wtss::WTSS("http://www.esensing.dpi.inpe.br/wtss")
# retrieve a time series
ndvi_ts <- wtss::time_series(wtss_server, "MOD13Q1", attributes = c("ndvi"), 
                             latitude = -10.408, longitude = -53.495)
# plot the time series
plot(ndvi_ts)
}
}
\author{
Gilberto Camara
}
