% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss-package.R
\docType{package}
\name{wtss-package}
\alias{wtss-package}
\alias{_PACKAGE}
\alias{wtss}
\title{wtss}
\description{
An R client to the web time series service (WTSS)
}
\section{WTSS API}{


Implements an R interface to a web time series service (WTSS) 
that offers time series of remote sensing data using a simple API. 
A WTSS server takes as input an Earth observation data cube, 
that has a spatial and a temporal dimension 
and can be multidimensional in terms of its attributes.

The WTSS API has four commands: 
\itemize{
   \item `wtss`: given an URL, creates a connection to a WTSS service
   \item `list_coverages`: returns a list of coverages (cubes) available 
       in the WTSS server.
   \item `describe_coverage`: returns the metadata for a given coverage.
   \item `time_series`: returns a time series for a spatio-temporal location.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-sensing/wtss/}
  \item Report bugs at \url{https://github.com/e-sensing/wtss/issues}
}

}
\author{
\strong{Maintainer}: Gilberto Camara \email{gilberto.camara@inpe.br}

Authors:
\itemize{
  \item Gilberto Queiroz \email{gilberto.queiroz@inpe.br}
  \item Pedro Andrade \email{pedro.andrade@inpe.br}
  \item Luiz Fernando Assis \email{luizffga@dpi.inpe.br}
  \item Rolf Simoes \email{rolf.simoes@dpi.inpe.br}
  \item Alber Sanchez \email{alber.ipia@inpe.br}
  \item Victor Maus \email{vwmaus1@gmail.com}
}

}
