% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_config.R
\name{wtss_config}
\alias{wtss_config}
\title{Reads a configuration file and loads it in the main environment}
\usage{
wtss_config()
}
\value{
A list with the configuration parameters used by wtss.
}
\description{
Reads a user-specified configuration file, located in a "config.yml" file
in the working directory. If this file is not found, reads a default package configuration file.
By default, the wtss configuration file "config.yml" is located at the directory "extdata" of the
package. The configuration file is an YAML file that should provide at least the following parameters:

default:
   WTSS_server    : "http://www.dpi.inpe.br/tws/wtss"

To see the contents of the configuration file, please use \code{\link[wtss]{wtss_config_show}}.
}
\examples{
# create configurtion file
config_wtss <- wtss_config()
# show configuration file
wtss_config_show()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
