% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_config.R
\name{wtss_config_show}
\alias{wtss_config_show}
\title{Shows the contents of the wtss configuration file}
\usage{
wtss_config_show()
}
\value{
List with the configuration parameters used by wtss.
}
\description{
Displays the contents of wtss configuration file. For more details
on how to set the configuration file, please use \code{\link[wtss]{wtss_config}}.
}
\examples{
wtss_config_show()
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
