% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_time_series.R
\name{wtss_filter_dates}
\alias{wtss_filter_dates}
\title{Filter dates in WTSS time series}
\usage{
wtss_filter_dates(data, start_date, end_date)
}
\arguments{
\item{data}{A sits tibble with time series.}

\item{start_date}{Starting date to be filtered}

\item{end_date}{End date to be filtered}
}
\value{
A time series filtered according to the start and end dates.
}
\description{
Converts data from a sits tibble to a list of a zoo series.
}
\examples{
{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
# connect to a WTSS server
wtss <- wtss::WTSS("http://www.esensing.dpi.inpe.br/wtss/")
# retrieve a time series as a tibble
ts_wtss  <- wtss::time_series(wtss, "MOD13Q1", c("ndvi","evi"), 
                longitude = -45.00, latitude  = -12.00,
                start_date = "2000-02-18", end_date = "2016-12-18")
# filter the time series
ts_wtss <- wtss::wtss_filter_dates(ts_wtss, start_date = "2001-01-01", end_date = "2014-12-31")
}
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
