% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{wtss_name}
\alias{wtss_name}
\title{Get name attribute from wtss objects}
\usage{
wtss_name(x)
}
\arguments{
\item{x}{a `describe_coverage` and object}
}
\value{
...
}
\description{
Get name attribute from wtss objects
}
