% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_time_series.R
\name{wtss_to_ts}
\alias{wtss_to_ts}
\title{Export data to be used to the ts format}
\usage{
wtss_to_ts(data, band)
}
\arguments{
\item{data}{A sits tibble with time series.}

\item{band}{Name of the band to be exported}
}
\value{
A time series in the ts format.
}
\description{
Converts data from a sits tibble to a list of a zoo series.
}
\examples{
# read a tibble with 400 samples of Cerrado and 346 samples of Pasture
# connect to a WTSS server
wtss <- wtss::WTSS("http://www.esensing.dpi.inpe.br/wtss/")
# retrieve a time series
ts_wtss  <- wtss::time_series(wtss, "MOD13Q1", c("ndvi","evi"), 
                longitude = -45.00, latitude  = -12.00,
                start_date = "2000-02-18", end_date = "2016-12-18")
# convert to zoo
ts <- wtss::wtss_to_ts(ts_wtss)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
