% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the list of cubes from the URL server}
\usage{
describe_coverage(URL, name, .print = TRUE)
}
\arguments{
\item{URL}{URL of the server}

\item{name}{name of coverage}

\item{.print}{Print the coverage description}
}
\value{
tibble with coverage description
}
\description{
Contacts the WTSS server to describe one coverage
}
\examples{
\dontrun{
# Using external server 
describe_coverage("https://brazildatacube.dpi.inpe.br/wtss/", 
                  "LC8_30_16D_STK-1")
}
}
