% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the list of cubes from the URL server}
\usage{
describe_coverage(wtss.obj, coverages)
}
\arguments{
\item{wtss.obj}{A WTSS object}

\item{coverages}{A character vector of coverage names}
}
\value{
A WTSS object with with coverage information
}
\description{
Contacts the WTSS server to request information about one or more coverages
}
\examples{
wtss  <-  WTSS("http://www.esensing.dpi.inpe.br/wtss/")
wtss  <- describe_coverage(wtss, wtss$coverages[1])
}
