% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the list of cubes from the URL server}
\usage{
describe_coverage(wtss.obj, name, .print = TRUE)
}
\arguments{
\item{wtss.obj}{A WTSS object}

\item{name}{A character vector of coverage names}

\item{.print}{Print the coverage description}
}
\value{
NULL if fails, TRUE if works
}
\description{
Contacts the WTSS server to describe one  coverages
}
\examples{
wtss  <-  WTSS("http://www.esensing.dpi.inpe.br/wtss/")
describe_coverage(wtss, wtss$coverages[1])
}
