% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the list of cubes from the URL server}
\usage{
describe_coverage(URL, name)
}
\arguments{
\item{URL}{URL of the server.}

\item{name}{name of coverage.}
}
\value{
a \code{tibble} with coverage description.
}
\description{
Requests the WTSS server to describe one coverage.
}
\examples{
\dontrun{
describe_coverage("https://brazildatacube.dpi.inpe.br/wtss/",
                  "LC8_30_16D_STK-1")
}
}
