% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Retrieves the metadata from one coverage}
\usage{
describe_coverage(URL, name)
}
\arguments{
\item{URL}{a \code{character} with URL of the server.}

\item{name}{a \code{character} with the coverage name.}
}
\value{
a \code{tibble} with coverage description.
}
\description{
Requests the WTSS server to get the metadata of one coverage.
The metadata includes minimum, maximum and missing values. 
Among other information from the coverage.
}
\examples{
\dontrun{
describe_coverage("https://brazildatacube.dpi.inpe.br/dev/wtss/v2/",
                  "LC8_30_16D_STK-1")
}
}
