% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-api.R
\name{.wtss_coverage_description}
\alias{.wtss_coverage_description}
\title{Decodes the description from a WTSS coverage}
\usage{
.wtss_coverage_description(URL, parsed_response)
}
\arguments{
\item{URL}{URL of the coverage}

\item{cov}{coverage response provided by WTSS service}
}
\description{
creates a tibble to store the description of the WTSS coverage
}
