% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-funs.R
\name{.wtss_get_response}
\alias{.wtss_get_response}
\alias{.get_request}
\title{Post a response to the WTSS server}
\usage{
.get_request(request, ..., n_tries = 3)
}
\arguments{
\item{request}{valid request according to the WTSS protocol}

\item{...}{additional parameters that can be added in httr.}
}
\value{
response from the server
}
\description{
Sends a request to the WTSS server and gets a response
}
