% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss_coverage.R
\name{.wtss_guess_satellite}
\alias{.wtss_guess_satellite}
\title{Try a best guess for the type of sensor/satellite}
\usage{
.wtss_guess_satellite(xres)
}
\arguments{
\item{xres}{xres of the coverage}
}
\value{
Satellite sensor pair
}
\description{
Based on resolution, tries to guess what is the satellite.
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
