% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{.wtss_list_coverages}
\alias{.wtss_list_coverages}
\title{Retrieves the list of cubes from the URL server}
\usage{
.wtss_list_coverages(wtss.obj)
}
\arguments{
\item{wtss.obj}{WTSS object associated to the service}
}
\value{
updated WTSS object.
}
\description{
Uses the WTSS services to print information and save metadata about a
chosen cube.
}
