% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{list_coverages}
\alias{list_coverages}
\title{List the coverages available in the WTSS service}
\usage{
list_coverages(URL)
}
\arguments{
\item{URL}{a \code{character} with URL of the server.}
}
\value{
a \code{list} with the links and names of each coverage.
}
\description{
Lists the data products made available by the 
 WTSS service for extracting and summarizing time series.
}
\examples{
\dontrun{
list_coverages("https://brazildatacube.dpi.inpe.br/dev/wtss/v2/")
}
}
