% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\name{WTSS}
\alias{WTSS}
\title{Creates a WTSS object}
\usage{
WTSS(URL)
}
\arguments{
\item{URL}{URL of the service provider.}
}
\value{
R WTSS object associated to the service.
}
\description{
Uses the WTSS services to print information and save metadata about a
chosen cube.
}
\examples{
{
wtss <-  wtss::WTSS("http://www.esensing.dpi.inpe.br/wtss/")
}
}
